chore: moved FeedCard component
